fix BufferedIO when size is greater than what can be in a 32-bit integer

+++ net.lecousin.core/src/main/java/net/lecousin/framework/io/buffering/BufferedIO.java
@@ -114,7 +114,7 @@ public abstract class BufferedIO extends BufferingManaged {
 		synchronized (buffer) {
 			buffer.buffer = new byte[index == 0 ? firstBufferSize : bufferSize];
 			loading =
-				io.readFullyAsync(index == 0 ? 0 : (firstBufferSize + (index - 1) * bufferSize), ByteBuffer.wrap(buffer.buffer));
+				io.readFullyAsync(index == 0 ? 0L : (firstBufferSize + (index - 1) * (long)bufferSize), ByteBuffer.wrap(buffer.buffer));
 		}
 		operation(loading).listenInline(new Runnable() {
 			@Override
